Add Laser Beamer,ada node ui and repair many bugs,add main menu

+++ Assets/Scripts/Turret/ProjectileSeting/Bullet/BulletDamage.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class BulletDamage : MonoBehaviour
 {
-    public void Damage(GameObject target, int damage)
+    public void Damage(GameObject target, float damage)
     {
         TakeDamage enemyTakeDamage = target.GetComponent<TakeDamage>();
         enemyTakeDamage.Damage(damage);
